requirements.txt
tqdm==4.67.1
pytorch_wavelets==1.3.0
PyWavelets ==1.8.0
matplotlib==3.10.1
pytorch_msssim==1.0.0
psutil==6.0.0